fix: failed to relaunch app to update
