+++ electron/main.ts
@@ -1,7 +1,6 @@
 import { app, BrowserWindow } from 'electron'
 
 import { join, resolve } from 'path'
-import { execSync } from 'child_process'
 
 /**
  * Managers
@@ -27,6 +26,7 @@ import { setupReactDevTool } from './utils/dev'
 import log from 'electron-log'
 
 import { start } from 'cortexso'
+import { cortexCppPort, cortexJsPort, cortexHost, cleanCortexProcesses } from './utils/cortex'
 
 const preloadPath = join(__dirname, 'preload.js')
 const rendererPath = join(__dirname, '..', 'renderer')
@@ -57,10 +57,6 @@ log.info('Starting jan from main thread..')
 // replace all console.log to log
 Object.assign(console, log.functions)
 
-const cortexJsPort = 1338
-const cortexCppPort = 3940
-const host = '127.0.0.1'
-
 app
   .whenReady()
   .then(() => {
@@ -90,10 +86,9 @@ app
     }
   })
 
-  .then(() => killProcessesOnPort(cortexCppPort))
-  .then(() => killProcessesOnPort(cortexJsPort))
+  .then(() => cleanCortexProcesses())
   .then(() => {
-    start('jan', host, cortexJsPort, cortexCppPort, getJanDataFolderPath())
+    start('jan', cortexHost, cortexJsPort, cortexCppPort, getJanDataFolderPath())
   })
   .then(createUserSpace)
   .then(migrate)
@@ -126,20 +121,9 @@ app.once('quit', async () => {
 })
 
 app.once('window-all-closed', async () => {
-  await stopApiServer()
   cleanUpAndQuit()
 })
 
-async function stopApiServer() {
-  // this function is not meant to be success. It will throw an error.
-  try {
-    await fetch(`http://${host}:${cortexJsPort}/v1/system`, {
-      method: 'DELETE',
-    })
-  } catch (error) {
-    // do nothing
-  }
-}
 
 /**
  * Handles various IPC messages from the renderer process.
@@ -148,88 +132,6 @@ function handleIPCs() {
   // Inject core handlers for IPCs
   // Handle native IPCs
   handleAppIPCs()
-}
-
-function killProcessesOnPort(port: number): void {
-  try {
-    console.log(`Killing processes on port ${port}...`)
-    if (process.platform === 'win32') {
-      killProcessesOnWindowsPort(port)
-    } else {
-      killProcessesOnUnixPort(port)
-    }
-  } catch (error) {
-    console.error(
-      `Failed to kill process(es) on port ${port}: ${(error as Error).message}`
-    )
-  }
-}
-
-function killProcessesOnWindowsPort(port: number): void {
-  let result: string
-  try {
-    result = execSync(`netstat -ano | findstr :${port}`).toString()
-  } catch (error) {
-    console.log(`No processes found on port ${port}.`)
-    return
-  }
-
-  const lines = result.split('\n').filter(Boolean)
-
-  if (lines.length === 0) {
-    console.log(`No processes found on port ${port}.`)
-    return
-  }
-
-  const pids = lines
-    .map((line) => {
-      const parts = line.trim().split(/\s+/)
-      return parts[parts.length - 1]
-    })
-    .filter((pid): pid is string => Boolean(pid) && !isNaN(Number(pid)))
-
-  if (pids.length === 0) {
-    console.log(`No valid PIDs found for port ${port}.`)
-    return
-  }
-  const uniquePids = Array.from(new Set(pids))
-  console.log('uniquePids', uniquePids)
-
-  uniquePids.forEach((pid) => {
-    try {
-      execSync(`taskkill /PID ${pid} /F`)
-      console.log(
-        `Process with PID ${pid} on port ${port} has been terminated.`
-      )
-    } catch (error) {
-      console.error(
-        `Failed to kill process with PID ${pid}: ${(error as Error).message}`
-      )
-    }
-  })
-}
-
-function killProcessesOnUnixPort(port: number): void {
-  let pids: string[]
-
-  try {
-    pids = execSync(`lsof -ti tcp:${port}`)
-      .toString()
-      .trim()
-      .split('\n')
-      .filter(Boolean)
-  } catch (error) {
-    if ((error as { status?: number }).status === 1) {
-      console.log(`No processes found on port ${port}.`)
-      return
-    }
-    throw error // Re-throw if it's not the "no processes found" error
-  }
-
-  pids.forEach((pid) => {
-    process.kill(parseInt(pid), 'SIGTERM')
-    console.log(`Process with PID ${pid} on port ${port} has been terminated.`)
-  })
 }
 
 /**
