+++ electron/utils/clean.ts
@@ -1,7 +1,14 @@
 import { windowManager } from './../managers/window'
 import { app } from 'electron'
+import { cleanCortexProcesses, stopCortexApiServer } from './cortex'
 
-export function cleanUpAndQuit() {
+
+/**
+ * Clean up windows then quit
+ */
+export async function cleanUpAndQuit() {
   windowManager.cleanUp()
+  await stopCortexApiServer()
+  await cleanCortexProcesses()
   app.quit()
 }
